feat: create controller getTopUsers
	- create controller
	- create tests
	- write swagger on this controller
	- refill database
	- fix bags related changed data
	- fix swagger documentation

+++ src/test/resources/create-tables.sql
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS products (
     p_id	binary(16),
     product_name varchar(256) NOT NULL,
     description	text NOT NULL,
-    price	decimal NOT NULL,
+    price	double NOT NULL,
     available_quantity int NOT NULL,
     category varchar(256) NOT NULL,
     is_available boolean NOT NULL,
